Rename security_pin to security_pin_page
